v0.5.6 (#340)
* ✅(add): document workflow experience

* 🍪(chore): cleanup

* 🪛(fix): connect info regexes

* 🪛(fix): duplicate shipment

* 🍪(chore): cleanup

* 🪛(fix): shipment form dirty

+++ internal/core/domain/validation.go
@@ -61,7 +61,7 @@ func ValidateTemperature(value any) error {
 		return nil
 	}
 
-	if int16(temperature) > temperatureMax || int16(temperature) < temperatureMin {
+	if temperature > temperatureMax || temperature < temperatureMin {
 		return eris.New("Invalid temperature. Please provide a valid temperature")
 	}
 	return nil
